refactors inputs and updates int and double configuration entry
