+++ src/main/kotlin/eu/florian_fuhrmann/musictimedtriggers/gui/views/components/inputs/DoubleField.kt
@@ -1,4 +1,4 @@
-package eu.florian_fuhrmann.musictimedtriggers.gui.views.components
+package eu.florian_fuhrmann.musictimedtriggers.gui.views.components.inputs
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -8,6 +8,7 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
+import eu.florian_fuhrmann.musictimedtriggers.gui.views.components.SimpleIconButton
 import eu.florian_fuhrmann.musictimedtriggers.utils.icons.MttIcons
 import org.jetbrains.jewel.foundation.modifier.trackActivation
 import org.jetbrains.jewel.ui.Outline
@@ -15,32 +16,37 @@ import org.jetbrains.jewel.ui.component.Text
 import org.jetbrains.jewel.ui.component.TextField
 
 @Composable
-fun IntegerField(
-    initialValue: Int,
-    minValue: Int = Int.MIN_VALUE,
-    maxValue: Int = Int.MAX_VALUE,
+@Deprecated("Uses deprecated TextField API. Use NumberField instead.")
+fun DoubleField(
+    initialValue: Double,
+    minValue: Double = -Double.MAX_VALUE,
+    maxValue: Double = Double.MAX_VALUE,
     buttons: Boolean = true, // whether to show plus and minus buttons
-    buttonStep: Int = 1,
-    plus: (Int) -> Int = { it + buttonStep },
-    minus: (Int) -> Int = { it -buttonStep },
+    buttonStep: Double = 1.0,
+    plus: (Double) -> Double = { it + buttonStep},
+    minus: (Double) -> Double = { it - buttonStep },
     placeholder: String = "",
     enabled: Boolean = true,
-    onChange: (Int) -> Unit = {}, //should only be called with valid values (in range values)
+    onChange: (Double) -> Unit = {}, //should only be called with valid values (in range values)
     onChangeInvalid: () -> Unit = {}, //will be called when the current input changes but is invalid
     outline: Outline = Outline.None
-): (Int) -> Unit {
+): (Double) -> Unit {
     var valueString: String by remember { mutableStateOf(initialValue.toString()) }
-    var intValue by remember { mutableStateOf(initialValue) }
-    val inRange: Boolean by remember { derivedStateOf { intValue in minValue..maxValue } }
-    var validInt: Boolean by remember { mutableStateOf(true) }
+    var doubleValue: Double by remember { mutableStateOf(initialValue) }
+    val inRange: Boolean by remember { derivedStateOf { doubleValue in minValue..maxValue } }
+    var validDouble: Boolean by remember { mutableStateOf(true) }
 
     Column {
         Row {
             TextField(
                 value = valueString,
                 onValueChange = {
-                    //only allow digits and minus
-                    if(it.any { c -> !c.isDigit() && c != '-' }) {
+                    //only allow digits, dots and minus
+                    if(it.any { c -> !c.isDigit() && c != '.' && c != '-' }) {
+                        return@TextField //just don't change the string value
+                    }
+                    //only allow one dot
+                    if(it.count { c -> c == '.' } > 1) {
                         return@TextField //just don't change the string value
                     }
                     //only allow minus as first char (check if any char after the first char is minus)
@@ -49,63 +55,61 @@ fun IntegerField(
                     }
                     //update string
                     valueString = it
-                    //get new int value
-                    val newIntValue = valueString.toIntOrNull()
-                    //check if the string could be parsed as a int
-                    if(newIntValue != null) {
-                        //if so mark as valid
-                        validInt = true
+                    //get double value
+                    val newDoubleValue = it.toDoubleOrNull()
+                    //check if the string could be parsed as a double
+                    if(newDoubleValue != null) {
+                        validDouble = true
                         //and update value
-                        intValue = newIntValue
+                        doubleValue = newDoubleValue
                         //call onChange if in range
                         if(inRange) {
-                            onChange(intValue)
+                            onChange(doubleValue)
                         } else {
                             //else call on chane invalid
                             onChangeInvalid()
                         }
                     } else {
-                        //if so mark as not valid
-                        validInt = false
+                        //the string is no valid double
+                        validDouble = false
+                        //call on change in valid
                         onChangeInvalid()
                     }
                 },
                 placeholder = {
                     Text(placeholder)
                 },
-                outline = outline,
+                outline = if(inRange && validDouble) { outline } else Outline.Error,
                 trailingIcon = {
                     if(buttons) {
                         Row {
                             SimpleIconButton(
-                                enabled = validInt && intValue > minValue,
+                                enabled = validDouble && doubleValue > minValue,
                                 forceHoverHandCursor = true,
                                 iconKey = MttIcons.minus,
                                 onClick = {
-                                    if(validInt) {
-                                        val newValue = minus(intValue).coerceAtLeast(minValue).coerceAtMost(maxValue)
-                                        if (intValue != newValue) {
-                                            //update int and string value and call onChange
-                                            intValue = newValue
-                                            valueString = newValue.toString()
-                                            onChange(intValue)
+                                    if(validDouble) {
+                                        val newValue = minus(doubleValue).coerceAtLeast(minValue).coerceAtMost(maxValue)
+                                        if (doubleValue != newValue) {
+                                            doubleValue = newValue
+                                            valueString = doubleValue.toString()
+                                            onChange(doubleValue)
                                         }
                                     }
                                 },
                                 modifier = Modifier.size(24.dp)
                             )
                             SimpleIconButton(
-                                enabled = validInt && intValue < maxValue,
+                                enabled = validDouble && doubleValue < maxValue,
                                 forceHoverHandCursor = true,
                                 iconKey = MttIcons.plus,
                                 onClick = {
-                                    if(validInt) {
-                                        val newValue = plus(intValue).coerceAtLeast(minValue).coerceAtMost(maxValue)
-                                        if (intValue != newValue) {
-                                            //update int and string value and call onChange
-                                            intValue = newValue
-                                            valueString = newValue.toString()
-                                            onChange(intValue)
+                                    if(validDouble) {
+                                        val newValue = plus(doubleValue).coerceAtLeast(minValue).coerceAtMost(maxValue)
+                                        if (doubleValue != newValue) {
+                                            doubleValue = newValue
+                                            valueString = doubleValue.toString()
+                                            onChange(doubleValue)
                                         }
                                     }
                                 },
@@ -118,9 +122,9 @@ fun IntegerField(
                 modifier = Modifier.trackActivation().fillMaxWidth()
             )
         }
-        if(!validInt) {
+        if(!validDouble) {
             Row {
-                Text(color = Color.Red, text = "Invalid Integer Number")
+                Text(color = Color.Red, text = "Invalid Decimal Number")
             }
         } else if(!inRange) {
             Row {
@@ -128,9 +132,8 @@ fun IntegerField(
             }
         }
     }
-
     //Return Function to update the value from the outside
     return {
-        intValue = it
+        doubleValue = it
     }
 }
